Move controllers into features
