+++ Stratis.Bitcoin/Features/MemoryPool/MempoolController.cs
@@ -2,10 +2,10 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using NBitcoin;
-using Stratis.Bitcoin.Features.MemoryPool;
+using Stratis.Bitcoin.Features.RPC.Controllers;
 using Stratis.Bitcoin.Utilities;
 
-namespace Stratis.Bitcoin.Features.RPC.Controllers
+namespace Stratis.Bitcoin.Features.MemoryPool
 {
     public class MempoolController : BaseRPCController
     {
