+++ Stratis.Bitcoin/Features/Wallet/WalletRPCController.cs
@@ -5,9 +5,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using NBitcoin;
 using Stratis.Bitcoin.Features.Miner;
-using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Features.RPC;
+using Stratis.Bitcoin.Features.RPC.Controllers;
 
-namespace Stratis.Bitcoin.Features.RPC.Controllers
+namespace Stratis.Bitcoin.Features.Wallet
 {
     public class WalletRPCController : BaseRPCController
     {
